Fix overflow of range counts calculation

+++ iv/aero/range_builder.h
@@ -77,12 +77,12 @@ class RangeBuilder : private core::Noncopyable<RangeBuilder> {
       if ((current.second + 1) >= it->first) {
         current.second = std::max(current.second, it->second);
       } else {
-        *counts += current.second - current.first + 1;
+        *counts += static_cast<uint32_t>(current.second - current.first) + 1;
         ranges_.push_back(current);
         current = *it;
       }
     }
-    *counts += current.second - current.first + 1;
+    *counts += static_cast<uint32_t>(current.second - current.first) + 1;
     ranges_.push_back(current);
     return ranges_;
   }
